proper tidy of ~/.netrc
R shouldn't write to `~/`, but old GDAL can't set a different path.  If we must write to ~/.netrc, we make sure we append to an existing ~/.netrc, make sure we append the same entry only once.  If no ~/.netrc exists, we make sure to create one as a symlink and remove it later.